Refactor code for running commands under docker into a module
This starts to cleanly separate the high-level concerns of the "build"
command from the implementation details of the environment in which
snakemake is run.  It also provides a foundation for sharing the docker
runner with other commands, like the "view" command I'm about to add.

The idea is that eventually there will be other runner modules that the
"build" command can use seamlessly, such as runner.local or
runner.singularity.