Add a deploy command for uploading data files
This replaces augur/scripts/s3.py for pushing files to staging or live.
It supports arbitrary S3 buckets and automatically detects and
invalidates CloudFront distributions for those buckets.

The design of the command allows for pluggable backend modules to
support other destinations in the future (like SFTP or other cloud
hosting services).