Check for newer versions of the CLI itself…
…in the check-setup and update commands.  These two commands already
come with the expectation that self-tests will be run and network
connections made, so it seems not very intrusive to do our own version
checking as well.

The idea is to prompt people to upgrade the CLI itself when they're
already thinking about their setup and not to distract or nag them when
they're running builds or viewing results.  Having people stay
relatively up-to-date on the CLI versions should make our development
easier.

Resolves #16.