mypy configuration for static type checking and error detection
The recently-merged deploy command includes Python 3.5+ compatible type
annotations in the nextstrain.cli.deploy.s3 module.  Types can be
checked by the mypy program, run as:

    mypy nextstrain

The minimal mypy.ini configuration file allows it to run clean.

The goal is to gradually add type annotations to the rest of the code so
we can catch errors earlier and be explicit about our interfaces.  This
matches well with the functional approach taken by the package.

I use a small, custom wrapper script for vim's quickfix mode to look at
mypy output, but there are many editor integrations for mypy, including
plugins for vim, Atom, and PyCharm.