runner.docker: Prune our old images after pull in update()
Otherwise the previous images are left around, using up potentially
quite a lot of disk space in a not-very-visible way.

Note that this requires labeled versions of nextstrain/base, so images
before that point won't ever be automatically pruned.  They have to be
manually cleaned up as a one-time step.

One downside of this is that locally built images which are only tagged
with "latest" will be removed when `nextstrain update` is run.
Protecting them requires tagging them with an additional value by
running `./devel/tag local-dev` in the docker-base repo, for example.

Resolves #23.