runner.docker: Remove per-command list of volumes
Upon further reflection, the per-command volumes logic unnecessarily
coupled the command and runner interfaces.  Its only tangible benefit
was reducing the number of possible options listed in `nextstrain view
--help`.  (This could also be seen as confusing, since the development
options section was present in other commands but not consistent across
them.)

The runner now allows overlaying of all defined components regardless of
command.  It may not make sense to use --augur with `nextstrain view`,
but it isn't prohibited.

This makes the way easier for a "native" runner.