Add a "native" runner which uses the ambient environment
This allows `nextstrain build --native` to work within an existing
Nextstrain environment, whether that's provided by conda or a cluster
module system or custom installs or some other thing.  The idea is that
the cli as a user-facing tool for Nextstrain is separate from a
containerized Nextstrain environment (although the two work well
together).

I chose the name "native" instead of "local" since the Docker containers
in our use case are also running on the local machine.  There will be
"remote" runners in the future, like AWS Batch, and I didn't want to
overload the term "local".

Resolves #1.