Run builds on AWS Batch
This adds an --aws-batch option to the "build" command which launches
the pathogen build as a job on AWS Batch.  It uploads the build
directory to S3, submits the job, monitors the job status, streams the
job logs, and downloads build results.

The UX aims to be very similar to that of local builds (either
containerized or native), so the "build" command stays in the foreground
and result files are written back directly to the local build directory.

AWS resources must be pre-configured and these are checked for (at least
at the surface level) by the "check-setup" command.  Documentation (and
possibly a setup script) for the required resources is still to come.

Two notable, independent improvements which are well-scoped and I think
worthwhile to do in the near future are:

  • better secured environment variable forwarding

  • unattended (background) builds whose results can be fetched at any
    later point

Comments on each of these are embedded in the relevant places in the
code.

The ability to overlay specific versions of Nextstrain components (like
--augur for the Docker runner) is another development-focused
improvement desired for the future.  However, it is highly co-dependent
on first implementing the more general case of remote build directories
(e.g.  `nextstrain build git+https://github.com/…`).

The new nextstrain.cli.hostenv module exists to consolidate local host
ambient environment handling for the Docker and AWS Batch runners.