runner.docker: Pull down new images by their build-* tag by default
Instead of tracking the mutable "latest" tag, `nextstrain update` now
looks up the latest image with a "build-*" tag and downloads that.  Our
build tags are, most importantly, not updated after creation and thus
are suitable references for reproducible runs.  The same is not true of
"latest".  For example, the output of `nextstrain version --verbose` now
includes the specific build tag.

Old images tagged with "build-*" are considered dangling and are removed
locally during the pruning process after update.

The current image in use is saved in the config file so that subsequent
`nextstrain build` runs can look it up there.  Only `nextstrain update`
will bump the image version, preserving that existing behaviour.

Setting the config file value to an image tagged something other than
"latest" or "build-*" tag will cause it to be preserved as-is, under the
assumption that it is some other mutable tag that updates in-place.
This works well for our "branch-*" tags, for example, and gracefully
handles forks of the nextstrain/base image too.