runner.aws_batch: Gracefully handle a missing log stream
The AWS Batch runner prints all log entries when a job transitions to a
terminal status and the log watcher hasn't started yet.  That handles
the case of a job failing or succeeding very quickly, before the runner
has time to start the log watcher.

If a job is externally canceled (e.g. via the AWS CLI or web console)
before it starts, however, it will never have a log stream associated
with it and requests to fetch a stream will fail.  Handle this edge case
by returning an empty generator from the JobState.log_entries() method.
This allows the runner to be mostly unaware of the details of the log
stream.

The log_watcher() method currently never triggers this edge case because
it is only invoked after the job has started and is guaranteed to have a
log stream.  An assertion is added to catch future missteps, however.