runner.aws_batch: Cancel submitted jobs cleanly on the first ^C
Exits immediately if a second ^C is received.  This convenience feature
is nice because it means jobs won't be orphaned to run for potentially a
long time (until the configured timeout in Batch).

The log watcher thread is daemonized so that the whole program exits
when an exception is raised in the main thread (e.g. the second
KeyboardInterrupt is re-raised).

Diff best viewed with whitespace ignored, as most of the job monitoring
code is indented further but otherwise unchanged.