Invoke pip using the name of the current python executable in upgrade instructions
On many Linux and macOS systems with dual versions of Python installed,
python is 2.7 and python3 is 3.x.  On other systems, python is 3.x and
python3 does not exist (notably conda environments and Windows).

This naming problem cascades to pip vs. pip3 as well.

As a remedy, suggest a pip upgrade invocation that runs pip as a module
using the current Python interpreter's executable if available, falling
back to the first python3 or python executable on PATH.