check-setup: Test that the configured Docker image is new enough
A minimum bound on the image build-* tag for what the CLI requires.  The
goal is to keep the image and CLI relatively loosely coupled so that
there aren't tight version dependencies between the two.  This generally
makes the user experience smoother and less of a hassle.  It also allows
users to stick to an older image version for longer even as they upgrade
the CLI, increasing reproducibility.  If we manage to keep them loosely
coupled successfully, then this minimum bound should require only
infrequent updates.