Work around mypy bug related to namespace packages
nextstrain is a namespace package (no __init__.py), which mypy doesn't
yet support in imports if also named on the command-line.¹  While the
previous version of this code was correct and ran just fine, mypy threw
an error.

The namespace package is a) more modern Python 3 style and b) better for
future co-operation in the nextstrain.* namespace, so I wanted to keep
it instead of caving and adding an __init__.py just for mypy.  Instead,
import a little more messily from nextstrain.cli.__version__.

¹ See <https://github.com/python/mypy/issues/5759>.