Declare missing dep on setuptools for pkg_resources.parse_version()
While setuptools is nearly guaranteed to be present, it's not totally
guaranteed.  One such example is when trying to build in a totally clean
virtual env with no site packages inherited from the system Python.

The version requirement marks the last parse_version() bug fix noted in
the setuptools release notes and dates to 2014, so shouldn't be an issue
for anyone.