runner.aws_batch: Stop deleting the build dir zip file from S3 and log stream from CloudWatch
While I expected folks to generally not interact with the AWS console or
command-line client to inspect these, that's turned out not to be the
case for some parallel builds of seasonal flu.

Our S3 bucket and job log group both have 30 day retention policies
applied, so this shouldn't cause the bucket and log group to bloat
beyond a month's worth of job data (which is the minimum duration we're
charged anyway).

Our documentation for AWS batch support recommends such retention
policies already, but I've updated the language to make it clear that it
really is a requirement.

Resolves #40.