runner.aws_batch: Restore mtimes of files when unzipping results
The dependency graph built by Snakemake relies on file modification
times to determine when targets need to be re-generated.  Previously all
mtimes were set to the time of extraction, which caused rules to
needlessly re-run when combining AWS Batch builds with local builds.

Note that because ZIP files use local timestamps for ZIP members but
don't record the local timezone (!), this code must assume the timezone
in which the ZIP file was created.  Our nextstrain/base container uses
GMT/UTC by default, so we can more or less safely assume that.  A future
improvement might be to switch to bzipped tarballs since tar guarantees
UTC internally.

Resolves #41.