runner.aws_batch: Correct typing of cpus and memory function params
These may legitimately be passed None, so their type must be wrapped in
Optional[] (which is equivalent to Union[x, None]).

Note that our type checking with mypy currently doesn't catch this
because some functions further up the call chain are still untyped.  I
opened issue #44 to address this in future work.