build: Detach from and re-attach to builds on AWS Batch
This adds a feature we've wanted from the beginning of the CLI.  By
starting the build with --detach, the job is only submitted and then the
CLI prints the command necessary to --attach to the job later.  This
command includes the job id and can be used as many times as desired.
It will even work on other computers or for other people, although you
will need to modify the local build path to a directory of your
choosing.

Jobs may also be interactively detached by pressing Control-Z.  Normally
this would suspend a Unix process (which could be restarted with `fg` or
`bg` or SIGCONT).  I think it's in the same spirit of suspending a local
process to detach from the remote job instead.  It also parallels nicely
with our existing Control-C job cancellation support.

There are currently no facilities to track job state locally or list
outstanding jobs, but these features could be added later if it seems
they'd be useful.  As it stands with this commit, a nice pattern for
launching multiple detached jobs and picking them up later is:

    # Loop over several `nextstrain build` commands, appending the last
    # line to a shell script.
    nextstrain build --aws-batch --detach build-a/ | tail -n1 | tee -a pickup-jobs.sh
    nextstrain build --aws-batch --detach build-b/ | tail -n1 | tee -a pickup-jobs.sh
    …

    # Then, sometime later:
    bash pickup-jobs.sh

Some existing code was further indented, so this diff is bested viewed
with --ignore-all-space.

Resolves #47.