gzip: Fix bug causing the EOF value (b'') to be returned too early
Depending on the data and conditions of compression, the .compress()
method will sometimes internally buffer all the data fed to it and
return only an empty byte string.  When that happened, the previous
implementation of GzipCompressingReader() mistook it as an EOF value,
stopped compression, and signaled EOF itself.  This led to some files,
but not all, being truncated during deploy/remote upload.

GzipCompressingReader() now uses its own internal buffer to make sure it
always has something to return to callers, even if multiple calls to
.compress() are necessary, thus avoiding a false EOF condition.  Support
for .read1() is dropped since it is impossible to implement its
behaviour given the API of gzip.  It was not used anyway.

Resolves #62.