runner: Don't use the default exec args if the program was changed
Previously using

    nextstrain build --exec foo <dir> ...

would result in

    foo --printshellcmds ...

being run since the default exec command is

    snakemake --printshellcmds ...

This fixes that.

Whether or not the CLI command accepts additional arguments or not is
still tied to the default exec command, which is a little weird, but
fine for now.  The --exec option exists for debugging and exceptional
use cases; we're not trying to do general-purpose command-running.