runner.aws_batch: Include the reason the container exited, if available
The container's exit reason is separate from the job status reason, and
it sometimes contains useful information like "OutOfMemoryError:
Container killed due to memory usage".  It only seems to exist in
certain cases—I can't find documentation listing when—but this code
tries to return it whenever it is available.