runner.aws_batch: Make the logging thread more resilient
On occasion, there would be ClientErrors caused by
InvalidSignatureException.  This is likely caused by laggy networking (I
was uploading a dataset to S3 simultaneously).  This crashes the logging
thread.

This change counts the number of successful and failed requests, and if
there have been successful requests, it ignores the failed requests.  If
there have never been any successful requests, then it raises an
exception after 10 failed requests.

Based on PR #79.

Co-authored-by: Thomas Sibley <tsibley@fredhutch.org>