Add retries for the S3 resource
This adds 3 retries.  I'm not sure if this is excessive, but I do see sporadic failures in uploading the zip.  The better solution is to do a multipart upload with retries (since it wouldn't need to start all over), but this should be a UX improvement.

Fixes https://github.com/nextstrain/cli/issues/75