version: Report versions from the native runner too
Restructures the way versions are reported by each runner to support
centralized output, which lets the command itself have consistent
formatting control.

The components reported by the Docker runner are re-ordered to put the
most common ones first, which happens to also be alphabetical.  This
affects the order of development flags --augur, --auspice, etc. in the
--help-all output, but not in any bad way.