Add support for commands to provide CPU and memory hints to runners
These hints can be used to both request enough resources when they're
dynamically available, such as AWS Batch jobs, as well as limit the
resources available to a command, such as with Docker.

As these will be command-dependent and not applicable to all commands,
it seems best to make them internal arguments rather than global runner
CLI options (like --exec).  Commands can expose CLI options as
appropriate, such as I'll do with the build command in the next commit.