Upload individual output files from AWS Batch job
This adds uploading of the individual files in the AWS Batch job working directory.

This is handy for situations when only one file needs to be downloaded, without downloading of the whole archive, which can be several gigabytes in size.

Note that this does not remove the upload of the large zip. Mostly because I simply don't know how it is used and whether there are tools that rely on its presence.

Implementation is a heavy copy-pate of the existing `upload_workdir()` function
https://github.com/nextstrain/cli/blob/103d3a3a1c890a7f574f8031cead3284697d24f7/nextstrain/cli/runner/aws_batch/s3.py#L39-L73
with the loop logic inverted: instead of add each file to zip, we create a new zip for each file