build: Support for selective downloading of remote build results
Adds two new options to the `nextstrain build` command:

  --download <pattern>
  --no-download

The former may be given multiple times and specifies patterns to match
against build dir files which were modified by the remote build.  The
latter skips downloading results entirely, which is useful if all you
care about are the logs (such as when re-attaching to a build or when a
build uploads results itself elsewhere).  The default is still to
download every modified file.

Currently limited to --aws-batch builds, as the only remote environment
supported.

This functionality will particularly help the ncov build shepherds, who
often have the need to download only a few files from a very large
build.

Resolves #104.  See also #83.