Bump minimum Python version from 3.5 to 3.6
Motivated by an upcoming dependency on fsspec which is 3.6+ and which
would incur more overhead than I'd like to make optional.

There are also many new standard library and language features starting
with 3.6, which would be nice to avail ourselves of in this codebase!

My original reason (back in 2018) for supporting 3.5 was that Ubuntu
16.04 LTS (Xenial) was only 2 years old, so pretty common still, and
shipped with 3.5.  It's now ~5 years old and almost out of the standard
LTS window.  The following LTS release, 18.04, shipped with 3.6.

A few places in code which handled 3.5 vs. 3.6+ differences still
remain, as they're not enough overhead to rip out (yet).

This change warrants a major version bump for the first release
including it.