Login and logout commands for authenticating with nextstrain.org
Username and password credentials are prompted for and exchanged for
authentication tokens which are then stored locally.  These tokens can
be periodically refreshed for ~30 days without requiring another
password prompt, similar to our nextstrain.org web session durations.
A new config file, ~/.nextstrain/secrets, is used to store the tokens so
that it may be protected by restricted file permissions, excluded from
backups/version control, etc. as suitable for each user.

There are three main parts to this functionality:

 1. The login and logout commands, which provide the user interface.
    These are responsible for prompting for credentials and outputting
    nice messaging to the user.

 2. An internal authentication library (nextstrain.cli.authn) used by
    the login/logout commands to make authentication attempts with
    nextstrain.org and transparently manage locally-saved tokens.  This
    library's API is designed to be used by other commands in the near
    future (e.g. "remote upload") which need to assert a user is logged
    in and get tokens for them.

 3. An internal AWS Cognito library that lightly wraps the boto3
    cognito-idp client and integrates an open-source implementation of
    the Secure Remote Password protocol used by Cognito.  It is intended
    only for use by our internal authn library.

First versions of this work used pyCognito in place of our own (3).
Upon closer inspection and consideration, it became clear to me that
pyCognito's approach is a bit sloppy and the API it provides a bit
awkward for our use case and also too easy to use incorrectly when
saving tokens locally.  I started to address some of this by submitting
patches upstream but, after more thinking about how pyCognito did much
more than we really needed, decided it'd be a better fit to implement
what became (2) and (3).  The APIs they expose and their conceptual
split support our use cases better and allow us more control.  If we
later find a better library for interacting with Cognito, we'll be able
to replace (3) with it pretty cleanly.  I did compare Warrant
(pyCognito's progenitor) to see if it was substantially different; it is
not.¹

The initial draft of this work in PR #92 was done by Kairsten Fay with
guidance from myself, but this final version ended up sharing little
with those initial drafts.  Some of the commits from her PR still remain
in this PR as early commits.

¹ https://github.com/tsibley/warrant-vs-pycognito