dev: Ignore a deprecation warning generated by flake8's usage of importlib_metadata
The warning was turned into an error by our pytest config, and thus
resulted in test failures.

importlib_metadata used to specially suppress this deprecation warning
for flake8, but that special-casing was removed in June without any
update to flake8's to avoid the warning.  See

    https://github.com/python/importlib_metadata/pull/319

and the flake8 PRs it links to for more background.