dev: Add (optional) type checking with Pyright to our tests
Pyright supports some checks and type inference that Mypy does not.  In
particular, it allows modules to be implementations of Protocols.  As
Mypy also supports some features that Pyright does not, we'll just use
both.

However, Pyright is written in TypeScript so cannot be declared as a
normal development dependency.  To use it, it must be manually installed
as `pyright` (e.g. with `npm install -g` or `snap`) or `npx` must be
available to run it on-demand.  The latter is usually easiest, provided
you have a new-enough version of Node already installed (>=12).  The
automated test is skipped if neither `pyright` nor `npx` are available.

The Pyright configuration here is stricter than the default level
("basic") but more permissive than the optional "strict" level.  I
enabled what I could cleanly, without requiring large code changes.
It's easier to enable stricter rules early, before future code is
written, than trying to do so later and fixing a lot of retroactive
infractions.

pyright-status: FAILING
mypy-status: PASSING