dev: Satisfy a new complaint by pyright's optional member access check
Newer versions think __doc__ could be None, which would mean this code
could call None.strip(), which is invalid.

__doc__ obviously can't be None because it's statically defined right
there in the docstring above (albeit implicitly by the language).

I don't like adding this sort of guard, but it's fine here or there so
long as false alarms like this don't metastasize into a larger issue.