Adjust deps to avoid incompatible versions of s3fs and aiobotocore
Affects fresh installs of nextstrain-cli with the latest aiobotocore
version, 2.0.0.

The latest aiobotocore, 2.0.0, does not fulfill the dep declarations of
the latest s3fs, 2021.10.1.  As a result, pip fulfills our s3fs dep with
2021.7.0 instead, which while declared to be compatible with
aiobotocore>=1.4.1 is not actually compatible in practice.  The result
is a runtime AttributeError on AioSession during upload to S3 in
`nextstrain build --aws-batch`.

When our previous deps on s3fs and aiobotocore[boto3] were introduced,
s3fs did not yet support its own [boto3] extra.  It does now, which
means we can let it declare the aiobotocore[boto3] dep itself and not
end up with conflicting dep declarations.  This results in fresh
installs using an older version of aiobotocore (1.4.x), but that's what
we want since it works.  Good to get the transitive dep out of our
declarations too.

Still a mess.  Dependencies are hard.

Resolves #133.