docs: Stop locking deps for Read The Docs and CI docs builds
Between issues with our custom theme (nextstrain-sphinx-theme) going
stale and Pipenv locking (which was the source of locking here), I think
this ended up causing more trouble than it potentially saved.

This may result in RTD build failures we wouldn't see otherwise, but
maybe better to know about those sooner than later anyway?  We can
always start locking again (maybe with the more minimal pip-compile
instead of pipenv) if this turns out to be a bad decision.  For
comparison, this is what we do for Augur's RTD build, and it seems to be
ok.