dev: Ditch Pipenv for a plain venv
Pipenv was nice at first for being an "all inclusive" tool, but with
time (mostly of it spent in Seattle Flu Study projects) it's shown its
warts with dependency-updating behaviour that's hard to reason about and
slowness or outright failure to resolve dependencies correctly.

In the latest indignity after today's change "Adjust deps to avoid
incompatible versions of s3fs and aiobotocore" (e9a9aa6), Pipenv was
failing to correctly resolve s3fs, aiobotocore, and boto3,
"successfully" producing:

  - a lock file that didn't include boto3
  - yet a venv that had boto3 installed
  - which when snapshot with `pip freeze` produced a requirements file
    uninstallable by modern pip.

I believe this is because Pipenv vendors an old version of pip and uses
an entirely different dependency resolution algorithm, but I didn't care
to debug it.

A plain venv is easier and more predictable to manage with that familiar
tool, pip.  It's also how our users will install nextstrain-cli, so it's
good to live in that same world.

This change means our dev envs no longer used locked deps… but that
seems fine.  It's how CI tests have been forever.  If we find we want
locking in the future, we can use the much more minimal pip-compile to
generate a dev-requirements.txt.