Add an rST → plain text Docutils writer extracted from Sphinx
Code does not compile yet.  This is a pure copy to make it easier to see
what small modifications are applied and for easier diffing later.

Sphinx is sprawling and unfortunately not suitable as a runtime dep for
the CLI, but Docutils (on which Sphinx is based) has no external deps
and is much smaller.  I'll use this extracted plain text writer with
Docutils to make a self-contained way to convert our command docstrings
from rST into both readable --help output _and_ HTML pages for
docs.nextstrain.org rather than maintain parallel doc.