Render URLs for most references as footnotes in --help output
Uses the standard Docutils target-notes directive.  Amends our
TextWriter to not wrap these generated footnotes so they remain
clickable and copy/paste-able.

Footnoted URLs are much easier to read than URLs inline with the text.
Note that bare URL references are still left inline and not moved into
footnotes, which is a limitation of the target-notes directive.