runner.aws_batch: Exit 1 if the job failed but there's no job exit code
Failed jobs do not have an exit code in Batch if the job failed for
infrastructural or reasons internal to AWS rather than job-specific
reasons.  For example, a job that fails because its EC2 instance is
shutdown in the middle of it will not have an exit code.

While `nextstrain build` used to exit 0 with success due to the missing
exit code, even though it correctly reported the job's failure and
failure reason.  Now it will exit 1 for a generic failure.