dev: Ignore deprecation warnings generated by pytest 7 for pytest-flake8
The warning was turned into an error by our pytest config, and thus
resulted in test failures.

There's an existing issue for pytest-flake8¹ that's 3 months old now
with no movement so far.  setuptools is also dealing with the same
warnings.²

If other issues arise that aren't fixed upstream, a longer term solution
could be doing what we do with mypy and simply invoke flake8 directly
instead of via a pytest plugin.

¹ https://github.com/tholo/pytest-flake8/issues/83
² https://github.com/pypa/setuptools/issues/3079