dev: Avoid pytest 7.0.0 on Python 3.9
It contains a fatal bug¹ when used on 3.9 because it incorrectly guards
a 3.10-only implementation detail with a >=3.9 version check, which
results in:

    ModuleNotFoundError: No module named 'importlib.readers'

when trying to load/parse Python modules.

Pytest 7.0.0 is only ~11 hours old at the moment, and another person
filed the upstream bug report as I was writing my own.  What a life,
living on the edge. 😬

Adjusts the warnings filter slightly in pytest.ini because
PytestRemovedIn8Warning is only present in pytest 7 and for now our
tests on 3.9 will use pytest 6.

¹ https://github.com/pytest-dev/pytest/issues/9608