runner.docker: Be more graceful of failure in the memory limit check
Don't assume that the container process will exit successfully.  For
example, on host systems with cgroups v2 (e.g. Ubuntu 22.04), the
cgroups v1 file /sys/fs/cgroup/memory/memory.limit_in_bytes doesn't
exist and so `cat` exits with error, which causes the container to exit
with error, which causes run_bash() to throw an exception.

This check is not critical and should be best effort.  If it fails
ungracefully, it impedes usage of the rest of the CLI.

Resolves #135.