dev: Run flake8 directly as a standalone test instead of using the pytest plugin
The pytest-flake8 plugin is not maintained and has been the source of a
lot of spurious CI failures as pytest's API changes and the plugin isn't
updated along with it.  Rather than continue to play whack-a-mole with
the plugin, just run flake8 ourselves as a single standalone test.  The
plugin broke out flake8 results per Python file in pytest's output, but
that's not really necessary.  This follows a similar move we made with
mypy some time ago.