Handle memory limits reported from Docker as floats
Fixes an issue on Mac OS X with Docker where `check-setup --set-default`
fails when the memory limits reported from the Docker OS are in floating
point values (e.g., "") that cannot be cast to integers by Python. This
commit addresses the issue by first casting the string to float before
casting to int.

Related to #153