runner.docker: Distinguish between an explicit vs. implicit "latest" image tag when updating
Behaviour with an implicit "latest" tag remains the same: it resolves to
the most recent "build-*" tag when `nextstrain update` is run which then
pins it in the config file.

An explicit "latest" tag used to do the same but now is preserved: it's
not updated to the most recent "build-*" tag during `nextstrain update`.
This allows folks to opt into tracking the leading edge of changes to
the image.

Resolves <https://github.com/nextstrain/cli/issues/96>.