runner.docker: Fix bug in old image pruning after update
Now prunes images older than the just-updated-to image, not the
just-updated-from image.  This bug meant that pruning would find older
images from *past* updates but skip the old image it just updated from,
i.e. if image N is the image just updated to, then images N-2 and older
were pruned instead of images N-1 and older.  An off-by-one bug, of a
sort.

I'd noticed this behaviour many times but had thought it was something
about the pruning causing other images to become prune-able (and thus
picked up in subsequent `nextstrain update` runs).  It was only when I
tried to fix that understanding of the bug with a loop and it didn't
work that I realized during debugging what was happening.