runner: Error instead of warn when --image is used with the "native" runner
It seems better to abort early with a clear error than to continue with
a warning that's easy to miss.  Particularly so since continuing while
ignoring --image might mean distant errors quite a while later (at best)
or hard-to-detect erroneous results (at worst).

Narrows the conditional to no longer exclude invocations where --image
was given explicitly but with a value that matched the default.