Build standalone installation archives with PyOxidizer
Produces .tar.gz (Linux, macOS) or .zip (Windows) archives with:

1. A "nextstrain" executable containing a bundled Python interpreter +
   stdlib + the Nextstrain CLI code + our dependencies.

2. External files (lib, data, etc) that are necessary but can't (for a
   variety of reasons) be bundled into the executable.

These installation archives can be downloaded, extracted, and run
in-place without even a Python interpreter being installed on the host
computer, hence the "standalone" moniker.

Currently we won't distribute these with/as releases, but I'd like to
get there and this is the first step.  I think good to be producing them
in CI even now so we can make sure they continue to work and can even
use them as unofficial "daily drivers" ourselves.