dev: Build dists explicitly as the first CI step
The test step now installs the built dists instead of implicitly
building by installing from the source repo.

This makes it possible to use the dists for other things later, like
automatic publishing to PyPI or building artifacts derived from the
dists.  It also brings testing a bit closer to the end user experience,
since the dists are what a user installs.