dev: Run source-based tests early in CI and separately from dist-based tests
Source-based tests are unit tests, type checking, linting, etc. which
need extra development dependencies and require a local source repo.

Dist-based tests are functional or integration tests which only exercise
the normal user interface (e.g. the `nextstrain` command) and do not
require a local source repo.

It's advantageous to run source-based tests early as if they fail then
we want to abort the rest of the CI run.  It's helpful to separate out
the two kinds of tests so that the local source repo and extra deps
needed by one don't potentially interfere with the other.