Suppress warning from cryptography 37 about deprecated Python 3.6 support
Version 37 still works on Python 3.6 but the deprecation causes loud,
user-visible warnings when running any command.

Version 38 will likely drop support, leaving 37 as the highest available
version on 3.6.  This is more reason to drop 3.6 ourselves eventually as
we'll not want to be using an older and older cryptography release.  In
the meantime, however, squash the warning since it does no good for end
users.  Frustratingly, cryptography appears to have intentionally made
this a UserWarning instead of a DeprecationWarning because they wanted
to force its visibility.