Suppress new warning from cryptography 39 about deprecated Python 3.6 support
Version 39 changed the wording of the deprecation warning added by
version 37.  39 still works on Python 3.6 but now warns that version 40
will be the last to include support.  Without supression, the
deprecation causes loud, user-visible warnings when running any command:

    Python 3.6 is no longer supported by the Python core team. Therefore,
    support for it is deprecated in cryptography. The next release of
    cryptography (40.0) will be the last to support Python 3.6.

This is more reason to drop 3.6 ourselves eventually as we'll not want
to be using an older and older cryptography release.  In the meantime,
however, squash the warning since it does no good for end users.

Resolves <https://github.com/nextstrain/cli/issues/245>.