runner.docker: Use os.getuid()/getgid() when present instead of conditioning on os.name
This is arguably more robust because it relies on "feature" detection
(of the kind popularized for browser web APIs) instead of checking a
proxy condition.  But the real reason for this is that Mypy can't/won't
narrow its attribute existence checking based on the os.name
conditional, so it produced errors on Windows where getuid() and
getgid() don't exist.