dev: Mark unreleased/in-development code with a +git local version part
The upshot of this is that any install from an unreleased version
(source repo, dist built by CI on branches, etc.) will be identifiable
as such and no longer confusable with the version of the last proper
release.  See PEP 440 § "Local version identifiers"¹ for details of this
syntax.

Going forward, the only commits where the version won't have this +git
local part will be commit/tag associated with an actual release.

¹ https://peps.python.org/pep-0440/#local-version-identifiers