dev: Avoid a confusing import binding situation
Import shadowing was happening here in the nextstrain.cli module because
of two subtly conflicting imports:

    import argparse
    ...
    from .argparse import ...

The first import binds the "argparse" symbol in nextstrain.cli to the
module from the stdlib, while the second rebinds "argparse" to the
nextstrain.cli.argparse module.¹

Later references in the same file to argparse.X worked because
nextstrain.cli.argparse happened to import (and not exclude from
re-export) the same referenced symbols.

This is confusing, so avoid it.

¹ See <https://docs.python.org/3/reference/import.html#submodules> for
  why.  Note that it only applies here because nextstrain.cli is a
  package with a nextstrain.cli.argparse submodule.  The same import
  statements in, say, nextstrain.cli.foo wouldn't conflict.