Detect when we're installed via a Conda package when suggesting an upgrade command
We prefer `mamba` over `conda` if the former is available on PATH and
detect the Conda environment prefix path as the closest directory to
ourselves containing a conda-meta/ directory.

It's important to use the Conda env prefix instead of env name since we
don't have reliable access to the name (e.g. when envs are stacked or
used without activation or created without a name in the first place).

By using the INSTALLER dist-info file, we correctly differentiate
between being installed via Pip into a Conda env (i.e. Conda is
incidental) and being installed via a Conda package into a Conda env
(i.e. Conda is instrumental).

Resolves <https://github.com/nextstrain/cli/issues/114>.